Move Consumption Budget tests
