+++ azurerm/internal/services/consumption/consumption_budget_resource_group_resource_test.go
@@ -1,4 +1,4 @@
-package tests
+package consumption
 
 import (
 	"fmt"
